feat: merged two test scenario since they are similar
